Move last remaining module out of src/sidebar/directive/test
Move Angular component testing utilities into
src/sidebar/components/test and rename the module to make it clear that
it contains Angular-specific code which we can delete once all
components are converted.
