+++ src/sidebar/components/test/angular-util.js
